Move RepoRefResolver to spi and remove imports from index
