+++ src/handlers/events/delivery/deploy/k8s/NoticeK8sProdDeployCompletion.ts
@@ -17,8 +17,8 @@
 import { EventFired, EventHandler, HandleEvent, HandlerContext, HandlerResult, logger, Secret, Secrets, Success } from "@atomist/automation-client";
 import { subscription } from "@atomist/automation-client/graph/graphQL";
 import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
-import { RepoRefResolver } from "../../../../../api/command/editor/support/RepoRefResolver";
 import { Goal } from "../../../../../api/goal/Goal";
+import { RepoRefResolver } from "../../../../../spi/repo-ref/RepoRefResolver";
 import { OnAParticularStatus } from "../../../../../typings/types";
 import { createStatus } from "../../../../../util/github/ghub";
 import { k8AutomationDeployContext, K8TargetBase } from "./RequestK8sDeploys";
