+++ src/handlers/events/delivery/goals/RespondOnGoalCompletion.ts
@@ -28,11 +28,11 @@ import {
     Value,
 } from "@atomist/automation-client";
 import { subscription } from "@atomist/automation-client/graph/graphQL";
-import { RepoRefResolver } from "../../../../api/command/editor/support/RepoRefResolver";
 import { addressChannelsFor } from "../../../../api/context/addressChannels";
 import { GoalCompletionListener, GoalCompletionListenerInvocation } from "../../../../api/listener/GoalsSetListener";
 import { SdmGoal } from "../../../../ingesters/sdmGoalIngester";
 import { fetchCommitForSdmGoal, fetchGoalsForCommit } from "../../../../internal/delivery/goals/support/fetchGoalsOnCommit";
+import { RepoRefResolver } from "../../../../spi/repo-ref/RepoRefResolver";
 import { OnAnyCompletedSdmGoal } from "../../../../typings/types";
 
 /**
