+++ src/handlers/events/delivery/verify/executeVerifyEndpoint.ts
@@ -15,12 +15,12 @@
  */
 
 import { sprintf } from "sprintf-js";
-import { RepoRefResolver } from "../../../../api/command/editor/support/RepoRefResolver";
 import { ExecuteGoalResult } from "../../../../api/goal/ExecuteGoalResult";
 import { ExecuteGoalWithLog, RunWithLogContext } from "../../../../api/goal/ExecuteGoalWithLog";
 import { Goal } from "../../../../api/goal/Goal";
 import { EndpointVerificationInvocation, EndpointVerificationListener } from "../../../../api/listener/EndpointVerificationListener";
 import { fetchGoalsForCommit } from "../../../../internal/delivery/goals/support/fetchGoalsOnCommit";
+import { RepoRefResolver } from "../../../../spi/repo-ref/RepoRefResolver";
 
 /**
  * What the SDM should define for each environment's verification
