+++ src/handlers/events/dry-run/OnDryRunBuildComplete.ts
@@ -19,7 +19,7 @@ import { subscription } from "@atomist/automation-client/graph/graphQL";
 import { GitHubRepoRef } from "@atomist/automation-client/operations/common/GitHubRepoRef";
 import { raiseIssue } from "@atomist/automation-client/util/gitHub";
 import { DryRunContext } from "../../../api/command/editor/dry-run/dryRunEditor";
-import { RepoRefResolver } from "../../../api/command/editor/support/RepoRefResolver";
+import { RepoRefResolver } from "../../../spi/repo-ref/RepoRefResolver";
 import { OnBuildCompleteForDryRun } from "../../../typings/types";
 import { createStatus } from "../../../util/github/ghub";
