+++ src/handlers/events/issue/NewIssueHandler.ts
@@ -16,9 +16,9 @@
 
 import { EventFired, EventHandler, HandleEvent, HandlerContext, HandlerResult, logger, Success } from "@atomist/automation-client";
 import { subscription } from "@atomist/automation-client/graph/graphQL";
-import { RepoRefResolver } from "../../../api/command/editor/support/RepoRefResolver";
 import { addressChannelsFor } from "../../../api/context/addressChannels";
 import { NewIssueListener, NewIssueListenerInvocation } from "../../../api/listener/NewIssueListener";
+import { RepoRefResolver } from "../../../spi/repo-ref/RepoRefResolver";
 import * as schema from "../../../typings/types";
 import { CredentialsResolver } from "../../common/CredentialsResolver";
