+++ src/handlers/events/repo/OnFirstPushToRepo.ts
@@ -18,9 +18,9 @@ import { EventFired, EventHandler, HandleEvent, HandlerContext, HandlerResult, l
 import { subscription } from "@atomist/automation-client/graph/graphQL";
 import { GitCommandGitProject } from "@atomist/automation-client/project/git/GitCommandGitProject";
 import * as _ from "lodash";
-import { RepoRefResolver } from "../../../api/command/editor/support/RepoRefResolver";
 import {AddressChannels, AddressNoChannels} from "../../../api/context/addressChannels";
 import { PushListener, PushListenerInvocation } from "../../../api/listener/PushListener";
+import { RepoRefResolver } from "../../../spi/repo-ref/RepoRefResolver";
 import * as schema from "../../../typings/types";
 import { CredentialsResolver } from "../../common/CredentialsResolver";
