+++ src/handlers/events/repo/OnPullRequest.ts
@@ -16,10 +16,10 @@
 
 import { EventFired, EventHandler, HandleEvent, HandlerContext, HandlerResult, Success } from "@atomist/automation-client";
 import { subscription } from "@atomist/automation-client/graph/graphQL";
-import { RepoRefResolver } from "../../../api/command/editor/support/RepoRefResolver";
 import { AddressChannels, addressChannelsFor } from "../../../api/context/addressChannels";
 import { PullRequestListener, PullRequestListenerInvocation } from "../../../api/listener/PullRequestListener";
 import { ProjectLoader } from "../../../spi/project/ProjectLoader";
+import { RepoRefResolver } from "../../../spi/repo-ref/RepoRefResolver";
 import * as schema from "../../../typings/types";
 import { CredentialsResolver } from "../../common/CredentialsResolver";
