+++ src/handlers/events/repo/OnUserJoiningChannel.ts
@@ -16,8 +16,8 @@
 
 import { EventFired, EventHandler, HandleEvent, HandlerContext, HandlerResult, Success } from "@atomist/automation-client";
 import { subscription } from "@atomist/automation-client/graph/graphQL";
-import { RepoRefResolver } from "../../../api/command/editor/support/RepoRefResolver";
 import { UserJoiningChannelListener, UserJoiningChannelListenerInvocation } from "../../../api/listener/UserJoiningChannelListener";
+import { RepoRefResolver } from "../../../spi/repo-ref/RepoRefResolver";
 import * as schema from "../../../typings/types";
 import { CredentialsResolver } from "../../common/CredentialsResolver";
