+++ src/index.ts
@@ -8,7 +8,6 @@ export * from "./api/command/editor/dry-run/dryRunEditor";
 export * from "./api/command/editor/support/allReposInTeam";
 export * from "./api/command/editor/support/confirmEditedness";
 export * from "./api/command/editor/support/editorWrappers";
-export * from "./api/command/editor/support/RepoRefResolver";
 export * from "./api/command/generator/BitBucketRepoCreationParameters";
 export * from "./api/command/generator/GeneratorConfig";
 export * from "./api/command/generator/SeedDrivenGeneratorParametersSupport";
@@ -60,6 +59,7 @@ export * from "./api/machine/GoalDrivenMachine";
 export * from "./api/machine/ListenerRegistration";
 export * from "./api/machine/MachineConfiguration";
 export * from "./api/machine/SoftwareDeliveryMachine";
+export * from "./api/machine/SoftwareDeliveryMachineOptions";
 export * from "./api/machine/wellKnownGoals";
 export * from "./api/machine/support/ListenerRegistrationSupport";
 export * from "./api/mapping/GoalSetter";
@@ -88,8 +88,6 @@ export * from "./api/registration/FingerprinterRegistration";
 export * from "./api/registration/PushReactionRegistration";
 export * from "./api/registration/ReviewerError";
 export * from "./api/registration/ReviewerRegistration";
-export * from "./api/repo-ref/AbstractRemoteRepoRef";
-export * from "./api/repo-ref/BitBucketServerRepoRef";
 export * from "./spi/artifact/ArtifactStore";
 export * from "./spi/build/Builder";
 export * from "./spi/deploy/Deployer";
@@ -98,4 +96,7 @@ export * from "./spi/deploy/Target";
 export * from "./spi/log/InterpretedLog";
 export * from "./spi/log/ProgressLog";
 export * from "./spi/project/ProjectLoader";
+export * from "./spi/repo-ref/AbstractRemoteRepoRef";
+export * from "./spi/repo-ref/BitBucketServerRepoRef";
+export * from "./spi/repo-ref/RepoRefResolver";
 export * from "./typings/types";
