+++ src/internal/delivery/deploy/executeDeploy.ts
@@ -21,9 +21,9 @@ import { ArtifactStore } from "../../../spi/artifact/ArtifactStore";
 import { checkOutArtifact, setEndpointGoalOnSuccessfulDeploy } from "./deploy";
 
 import * as _ from "lodash";
-import { RepoRefResolver } from "../../..";
 import { ExecuteGoalWithLog, RunWithLogContext } from "../../../api/goal/ExecuteGoalWithLog";
 import { Target } from "../../../spi/deploy/Target";
+import { RepoRefResolver } from "../../../spi/repo-ref/RepoRefResolver";
 
 /**
  * Execute deploy with the supplied deployer and target
